// interfaces/AccountDataInterface.ts
export interface AccountDataInterface {
  firstName: string;
  lastName: string;
  email: string;
  telephone: string;
  fax: string;
  company: string;
  address1: string;
  address2: string;
  city: string;
  country: string;
  zone: string;
  postcode: string;
  loginname: string;
  password: string;
  newsletter: boolean;
  agree: boolean;
}